requirements.txt
numpy==1.15.0
opencv-python==3.4.2.17
Pillow==5.2.0
PuLP==1.6.8
pyparsing==2.2.0
pypng==0.0.18
pytesseract==0.2.4
